Make use of longs instead of integers since large files have a size greater than 2147483648 bytes (2.1 gigabytes)

+++ TorSharp/TorrentFile/Bencoding/Lexer/Lexer.cs
@@ -27,7 +27,7 @@ public class Lexer : ILexer
             if (input[i] is 'e')
             {
                 i++;
-                return new IntegerToken { Value = int.Parse(result) };
+                return new IntegerToken { Value = long.Parse(result) };
             }
             result += input[i];
         }
